new colmap, support service and update docker compose

+++ run_service.py
@@ -21,21 +21,59 @@ def authenticate_drive():
     credentials = Credentials.from_service_account_file(CREDENTIALS_FILE, scopes=scopes)
     return build("drive", "v3", credentials=credentials)
 
+def check_n_frames(new_video_id):
+    file_name = new_video_id['name']
+    n = 100
+    
+    if '_n_' in file_name:
+        n_tmp = file_name.split('_n_')[1]
+        n = int(n_tmp.split('.')[0])
+
+    print(f"Using {n} frames")
+    return n
+
 # Download a file from Google Drive
 def download_file(drive_service, new_video_id, ws_path):
     file_id = new_video_id['id']
     file_name = new_video_id['name']
     local_path = os.path.join(ws_path, file_name)
+
+    if os.path.isfile(local_path):
+        print(f"{file_name} already downloaded")
+        return
+
     request = drive_service.files().get_media(fileId=file_id)
+    print(f"Doanloading {file_name} .....")
     with open(local_path, "wb") as f:
         f.write(request.execute())
 
+    print(f"{file_name} Downloaded!")
+
 # Upload a file to Google Drive
-def upload_to_drive(drive_service, ws_path, output_folder_id, renamed_file_name):
+def upload_to_drive(drive_service, ws_path, output_folder_id, renamed_file_name, retries=5):
+    renamed_file_name = renamed_file_name.split('.')[0] + '.ply'
     file_metadata = {"name": renamed_file_name, "parents": [output_folder_id]}
     local_file_path = os.path.join(ws_path, "model/point_cloud/iteration_7000/point_cloud.ply")
+    if not os.path.isfile(local_file_path):
+        print("Something went wrong, exit!")
+        sys.exit(1)
+
+    print("Uploading 3dgs ....") 
     media = MediaFileUpload(local_file_path, mimetype="application/octet-stream")
-    drive_service.files().create(body=file_metadata, media_body=media, fields="id").execute()
+    
+    for attempt in range(retries):
+        try:
+            drive_service.files().create(body=file_metadata, media_body=media, fields="id").execute()
+            print("3dgs uploaded!")
+            return
+            
+        except Exception as e:
+            print(f"Upload failed (attempt {attempt + 1}/{retries}): {e}")
+            if attempt < retries - 1:
+                time.sleep(5)  # Wait before retrying
+            else:
+                raise
+
 
 # Check for new videos in Google Drive
 def fetch_new_videos(drive_service):
@@ -69,8 +107,8 @@ def create_local_ws(new_video_id):
     return ws_path
 
 
-def create_gs(ws_path):
-    command = f"conda run --live-stream -n sugar python do_all.py -s {ws_path} -n 100"
+def create_gs(ws_path, n_frames=100):
+    command = f"conda run --live-stream -n sugar python do_all.py -s {ws_path} -n {n_frames}"
     subprocess.run(command, shell=True, check=True)
 
 
@@ -91,7 +129,8 @@ def main():
         print(f"Found new video: { new_video_id['name']}")
         ws_path = create_local_ws(new_video_id)
         download_file(drive_service, new_video_id, ws_path)
-        create_gs(ws_path)
+        n_frames = check_n_frames(new_video_id)
+        create_gs(ws_path, n_frames)
         upload_to_drive(drive_service, ws_path, OUTPUT_FOLDER_ID, new_video_id['name'])
         move_file(drive_service, new_video_id['id'], DONE_FOLDER_ID)
